Work in progress: adding project id tracking code
Initial changes to support project id based quota.
Adding functions to support setting a new project id and tracking it
incore, Currently adding a support with limit of a single project id.
for an inode. This would make it easier to incrementally develop the
feature. Will change this once the essential pieces for testing
are in.

Also, project based quota will be mutually exclusive to existing directory
quota. However currently keeping the code in existing quota and marker
translators and enable it using  volume option. I will move this out to a new
version of quota and marker after basic testing.

More details:
https://docs.google.com/document/d/1bXIOfAbad9mrezUXdlOsgrlvABImZyraXsEFBkW0wcc/edit